+++ R/xvars.R
@@ -35,22 +35,16 @@ make_decimal_date <- function(date, frequency) {
 get_x_values <- function(data, x) {
   outx <- list()
   for (p in names(data)) {
-    if (!is.null(x[[p]])) {
-      if (x[[p]] %in% names(data[[p]])) {
-        outx[[p]] <- data[[p]][[x[[p]]]]
-        # if is dates, convert to year fractions
-        if (lubridate::is.Date(outx[[p]]) || lubridate::is.POSIXt(outx[[p]])) {
-          freq <- frequencyof(outx[[p]])
-          outx[[p]] <- make_decimal_date(outx[[p]], freq)
-          # Add a little helper to tell other functions we have time series data
-          outx[[paste0(p, "ts")]] <- TRUE
-          outx[[paste0(p, "freq")]] <- freq
-        }
-      } else {
-        stop(paste0("The x variable you supplied for panel ", p, " (" , x[[p]], ") is not a variable in the data you supplied."))
+    if (x[[p]] %in% names(data[[p]])) {
+      outx[[p]] <- data[[p]][[x[[p]]]]
+      # if is dates, convert to year fractions
+      if (lubridate::is.Date(outx[[p]]) || lubridate::is.POSIXt(outx[[p]])) {
+        freq <- frequencyof(outx[[p]])
+        outx[[p]] <- make_decimal_date(outx[[p]], freq)
+        # Add a little helper to tell other functions we have time series data
+        outx[[paste0(p, "ts")]] <- TRUE
+        outx[[paste0(p, "freq")]] <- freq
       }
-    } else {
-      stop(paste0("Have not supplied an x variable for panel ", p, " and cannot guess it because it is not a time series."))
     }
   }
   return(outx)